move the VEH out of dllmain
