+++ Dalamud.Boot/veh.h
@@ -1,0 +1,8 @@
+#pragma once
+
+namespace veh
+{
+    bool add_handler();
+    bool remove_handler();
+    void* get_handle();
+}
